Make it more clear that should one line syntax is still good to use
Do not trust people to read three lines of text after the last code block in a section. Include this so that it will stand out that this is still supported. I skimmed over it and got confused why other sections were still using should. I didn't find the explanation about this until I went to the github discussion for the section. Don't trust people to notice stuff at the bottom of these sections.